+++ R/edit_matches.R
@@ -1,19 +1,24 @@
+
 #' Edit matches manually in a csv-table
 #'
 #' Allows the user to match concepts with an already existing ontology, without
 #' actually writing into the ontology, but instead storing the resulting
 #' matching table as csv.
 #' @param new [`data.frame(.)`][data.frame]\cr the new concepts that shall be
-#'   manually matched.
-#' @param target [`data.frame(.)`][data.frame]\cr the attributes, in terms of
-#'   columns in the ontology, of new concepts that help to match new and target
-#'   concepts manually.
+#'   manually matched, includes "label", "class" and "has_broader" columns.
+#' @param topLevel [`logical(1)`][logical]\cr whether or not the new concepts
+#'   are at the highest level only, i.e., have to be matched without context, or
+#'   whether they are contain columns that must be matched within parent
+#'   columns.
 #' @param source [`character(1)`][character]\cr any character uniquely
 #'   identifying the source dataset of the new concepts.
 #' @param ontology [`ontology(1)`][list]\cr either a path where the ontology is
 #'   stored, or an already loaded ontology.
 #' @param matchDir [`character(1)`][character]\cr the directory where to store
 #'   source-specific matching tables.
+#' @param stringdist [`logical(1)`][logical]\cr whether or not to use string
+#'   distance to find matches (should not be used for large datasets/when a
+#'   memory error is shown).
 #' @param verbose [`logical(1)`][logical]\cr whether or not to give detailed
 #'   information on the process of this function.
 #' @param beep [`integerish(1)`][integer]\cr Number specifying what sound to be
@@ -36,28 +41,32 @@
 #' @return A table that contains all new matches, or if none of the new concepts
 #'   weren't already in the ontology, a table of the already sucessful matches.
 #' @importFrom checkmate assertDataFrame assertNames assertCharacter
-#'   assertFileExists testFileExists
+#'   assertFileExists testFileExists assertDirectoryExists
 #' @importFrom utils tail head
-#' @importFrom stringr str_split
+#' @importFrom stringr str_split str_extract
 #' @importFrom readr read_csv write_csv cols
 #' @importFrom dplyr filter rename full_join mutate if_else select left_join
-#'   bind_rows distinct arrange any_of
+#'   bind_rows distinct arrange any_of if_any
 #' @importFrom tidyselect everything starts_with
-#' @importFrom tidyr pivot_longer pivot_wider
+#' @importFrom tidyr pivot_longer pivot_wider separate_wider_delim
+#'   separate_wider_regex
 #' @importFrom tibble add_column
 #' @importFrom fuzzyjoin stringdist_left_join
 #' @export
 
-edit_matches <- function(new, target = NULL, source = NULL,
-                         ontology = NULL, matchDir = NULL, verbose = TRUE,
+edit_matches <- function(new, topLevel, source = NULL, ontology = NULL,
+                         matchDir = NULL, stringdist = TRUE, verbose = TRUE,
                          beep = NULL){
 
-  assertCharacter(x = new)
-  assertNames(x = names(target), must.include = c("class", "has_broader"))
+  assertDataFrame(x = new)
+  assertNames(x = names(new), must.include = c("label", "class", "has_broader"))
+  assertLogical(x = topLevel, any.missing = FALSE, len = 1)
   assertCharacter(x = source, len = 1, any.missing = FALSE)
+  assertDirectoryExists(x = matchDir, access = "rw")
 
-  intPaths <- getOption(x = "adb_path")
-  sourceFile <- paste0("match_", source, ".csv")
+  sourceFile <- paste0("match_", source, ".rds")
+  sourceID <- get_source(label = source, ontology = ontology) %>%
+    pull(id)
 
   if(inherits(x = ontology, what = "onto")){
     ontoPath <- NULL
@@ -72,83 +81,107 @@ edit_matches <- function(new, target = NULL, source = NULL,
 
   # get the classes within which to search
   filterClasses <- ontology@classes$harmonised %>%
-    filter(label %in% target$class)
-  filterClassLevel <- length(str_split(string = filterClasses$id, pattern = "[.]")[[1]])
-  if(dim(filterClasses)[1] == 0){
-    stop("no classes are matched in the ontology.")
-  }
-  while(!any(is.na(filterClasses$has_broader))){
-    filterClasses <- ontology@classes$harmonised %>%
-      filter(label %in% filterClasses$label | label %in% filterClasses$has_broader)
-  }
-  filterClasses <- filterClasses %>%
-    pull(label) %>%
-    unique()
+    filter(label %in% new$class)
 
-  #
-  target <- target %>%
-    mutate(lvl = length(str_split(has_broader, "[.]")[[1]]))
-
-  if(all(target$lvl < filterClassLevel-1)){
-    parentFilter <- unique(target$has_broader)
-    withBroader <- NULL
+  if(topLevel){
+    joinCols <- c("label", "class")
   } else {
-    parentFilter <- NULL
-    withBroader <- "has_broader"
+    joinCols <- c("label", "class", "has_broader")
   }
 
-  temp <- get_concept(label = new, class = target$class, has_broader = target$has_broader, ontology = ontology) %>%
-    left_join(tibble(label = new, has_broader = target$has_broader), ., by = c("label", "has_broader"))
+  if(dim(filterClasses)[1] != 0){
 
-  # determine previous matches
-  if(testFileExists(paste0(matchDir, sourceFile))){
-    prevAvail <- TRUE
-    prevMatches <- read_csv(paste0(matchDir, sourceFile), col_types = cols(.default = "c"))
-
-    prevMatchLabels <- prevMatches %>%
-      filter(class %in% filterClasses) %>%
-      pivot_longer(cols = c(has_broader_match, has_close_match, has_exact_match, has_narrower_match), values_to = "labels") %>%
-      filter(!is.na(labels)) %>%
-      distinct(labels) %>%
-      separate_rows(labels, sep = " \\| ") %>%
-      pull(labels)
-  } else {
-    prevAvail <- FALSE
-    # if no previous matches are present, match the new concepts with the already
-    # harmonised concepts in assumption that a match on the term is also a match on
-    # the underlying concept
-    tempMatch <- temp %>%
-      mutate(has_close_match = label) %>%
-      select(label, has_close_match, has_broader)
-    prevMatches <- ontology@concepts$harmonised %>%
-      filter(class %in% filterClasses) %>%
-      select(-has_close_match) %>%
-      left_join(tempMatch, ., by = c("label", "has_broader")) %>%
-      filter(!is.na(id)) %>%
-      mutate(has_broader_match = NA_character_,
-             has_exact_match = NA_character_,
-             has_narrower_match = NA_character_)
-
-    prevMatchLabels <- prevMatches %>%
-      distinct(label) %>%
-      separate_rows(label, sep = " \\| ") %>%
-      pull(label)
-
-    if(dim(prevMatches)[1] == 0){
-      prevMatches[1,] <- "ignore"
-      prevMatches$class <- filterClasses[1]
+    filterClassLevel <- length(str_split(string = filterClasses$id, pattern = "[.]")[[1]])
+    if(dim(filterClasses)[1] == 0){
+      stop("no classes are matched in the ontology.")
     }
+    while(!any(is.na(filterClasses$has_broader))){
+      filterClasses <- ontology@classes$harmonised %>%
+        filter(label %in% filterClasses$label | label %in% filterClasses$has_broader)
+    }
+    filterClasses <- filterClasses %>%
+      pull(label) %>%
+      unique()
+
+    # identify level of the target concepts
+    new <- new %>%
+      mutate(lvl = length(str_split(has_broader, "[.]")[[1]]))
+
+    if(any(ontology@classes$harmonised$label %in% getOption("gazetteer_top"))){
+      topLength <- length(str_split(string = ontology@classes$harmonised |> filter(label %in% getOption("gazetteer_top")) |> pull(id), pattern = "[.]")[[1]])
+      new <- new %>%
+        mutate(top = str_extract(string = has_broader, pattern = paste0("(.[[:digit:]]+){", topLength-1, "}")))
+
+      parentFilter <- unique(new$top)
+      withBroader <- NULL
+    } else {
+
+      # set a filter in case target concepts have broader concepts
+      if(all(new$lvl <= filterClassLevel-1)){
+        parentFilter <- unique(new$has_broader)
+        withBroader <- NULL
+      } else {
+        parentFilter <- NA
+        withBroader <- "has_broader"
+      }
+
+    }
+
+    ignoreClass <- tail(filterClasses, 1)
+  } else {
+    filterClasses <- get_class(ontology = ontology) %>%
+      pull(label)
+    parentFilter <- NA
+    withBroader <- NULL
+    ignoreClass <- head(filterClasses, 1)
   }
 
-  # gather all concepts for the focal data-series (previous matches from
-  # matching table and matches that may already be in the ontology) ...
-  dsConcepts <- prevMatches %>%
+  # ontoMatches <- get_concept(label = new$label, class = new$class, has_broader = new$has_broader, ontology = ontology) %>%
+  #   rename(harmLab = label) %>%
+  #   pivot_longer(cols = c(has_broader_match, has_close_match, has_exact_match, has_narrower_match),
+  #                names_to = "match", values_to = "label")
+
+  # determine previous matches from matching table (and make them long)
+  if(testFileExists(paste0(matchDir, sourceFile))){
+    dsMatches <- readRDS(file = paste0(matchDir, sourceFile))
+  } else {
+    dsMatches <- tibble(id = character(), label = character(), class = character(), has_broader = character(), description = character(),
+                        has_broader_match = character(), has_close_match = character(), has_exact_match = character(), has_narrower_match = character())
+  }
+
+  dsMatchesLong <- dsMatches %>%
     filter(class %in% filterClasses) %>%
     rename(harmLab = label) %>%
     pivot_longer(cols = c(has_broader_match, has_close_match, has_exact_match, has_narrower_match),
                  names_to = "match", values_to = "label") %>%
-    separate_rows(label, sep = " \\| ") %>%
-    full_join(temp, by = c("label", "class", "id", "has_broader", "description")) %>%
+    separate_longer_delim(cols = label, delim = " | ") |>
+    filter(!is.na(label))
+
+  # ignore concepts that were previously tagged 'ignore'
+  tempIgnore <- dsMatchesLong |>
+    filter(harmLab == "ignore") |>
+    select(label, class, harmLab) |>
+    distinct()
+
+  new <- new |>
+    left_join(tempIgnore, by = c("label", "class")) |>
+    filter(!harmLab %in% "ignore") |>
+    filter(label != "") |>
+    distinct() |>
+    select(-harmLab)
+
+  # ... and return an empty object if everything is to ignore
+  if(dim(new)[1] == 0){
+    out <- tibble(id = character(), has_broader = character(), label = character(), class = character(),
+                  description = character(), match = character(), external = character())
+    return(out)
+  }
+
+  # gather all concepts for the focal data-series (previous matches from
+  # matching table and matches that may already be in the ontology) and join
+  # with new concepts
+  dsConcepts <- dsMatchesLong %>%
+    full_join(new |> select(all_of(joinCols)), by = joinCols) %>%
     mutate(harmLab = if_else(is.na(harmLab), label, harmLab),
            label = if_else(is.na(match), if_else(!is.na(id), label, NA_character_), label),
            match = if_else(is.na(match), if_else(!is.na(id), "has_close_match", "sort_in"), match)) %>%
@@ -162,64 +195,102 @@ edit_matches <- function(new, target = NULL, source = NULL,
     dsConcepts <- dsConcepts %>%
       select(-sort_in)
   }
-  if(!"has_close_match" %in% colnames(dsConcepts)){
-    dsConcepts <- dsConcepts %>%
-      add_column(has_close_match = NA_character_, .after = "description")
-  }
   if(!"has_broader_match" %in% colnames(dsConcepts)){
     dsConcepts <- dsConcepts %>%
-      add_column(has_broader_match = NA_character_, .after = "has_close_match")
+      add_column(has_broader_match = NA_character_, .after = "description")
+  }
+  if(!"has_close_match" %in% colnames(dsConcepts)){
+    dsConcepts <- dsConcepts %>%
+      add_column(has_close_match = NA_character_, .after = "has_broader_match")
   }
   if(!"has_exact_match" %in% colnames(dsConcepts)){
     dsConcepts <- dsConcepts %>%
-      add_column(has_exact_match = NA_character_, .after = "has_broader_match")
+      add_column(has_exact_match = NA_character_, .after = "has_close_match")
   }
   if(!"has_narrower_match" %in% colnames(dsConcepts)){
     dsConcepts <- dsConcepts %>%
       add_column(has_narrower_match = NA_character_, .after = "has_exact_match")
   }
+  dsConcepts <- dsConcepts |>
+    select(label, class, id, has_broader, description, has_broader_match, has_close_match, has_exact_match, has_narrower_match)
 
   # ... and determine which are already included concepts and which are still missing
   inclConcepts <- dsConcepts %>%
     filter(!is.na(id))
   missingConcepts <- dsConcepts %>%
-    filter(is.na(id) & !label %in% prevMatchLabels)
+    filter(is.na(id))
 
   if(dim(missingConcepts)[1] != 0){
 
-    if(!is.null(parentFilter)){
-      if(!all(is.na(parentFilter))){
-        toRelate <- make_tree(id = parentFilter, ontology = ontology)
-      } else {
-        toRelate <- ontology@concepts$harmonised
-      }
-    } else {
-      toRelate <- ontology@concepts$harmonised
+    toRelate <- ontology@concepts$harmonised
+    if(!any(is.na(parentFilter))){
+      toRelate <- make_tree(id = parentFilter, ontology = ontology)
     }
+
+    extConcepts <- ontology@concepts$external %>%
+      separate_wider_delim(cols = id, names = c("dataseries", "nr"), delim = "_", cols_remove = FALSE) %>%
+      rowwise() %>%
+      mutate(label = paste0(label, " [", dataseries, "]")) %>%
+      select(external = label, temp = id)
+
     relate <- toRelate %>%
-      select(id, label, class, has_broader) %>%
       filter(class %in% filterClasses) %>%
-      left_join(inclConcepts, by = c("id", "label", "class", "has_broader")) %>%
+      filter(!label == "ignore") %>%
+      pivot_longer(cols = c(has_broader_match, has_close_match, has_exact_match, has_narrower_match),
+                   names_to = "match", values_to = "external") %>%
+      separate_rows(external, sep = " \\| ") %>%
+      separate_wider_delim(cols = external, names = c("temp"), delim = ".", too_many = "drop") %>%
+      left_join(extConcepts, by = "temp") %>%
+      group_by(across(all_of(c("id", "label", "class", "has_broader", "description", "match")))) %>%
+      summarise(external = paste0(na.omit(external), collapse = " | ")) %>%
+      ungroup() %>%
+      mutate(external = na_if(external, "")) %>%
+      pivot_wider(id_cols = c("id", "label", "class", "has_broader", "description"), names_from = match, values_from = external) %>%
+      rowwise() %>%
+      mutate(description = paste0(na.omit(c(description,
+                                            if_else(!is.na(has_close_match), paste0("close: ", has_close_match), NA),
+                                            if_else(!is.na(has_broader_match), paste0("broader: ", has_broader_match), NA),
+                                            if_else(!is.na(has_narrower_match), paste0("narrower: ", has_narrower_match), NA))),
+                                  collapse = " -- ")) %>%
+      mutate(description = na_if(description, "")) %>%
+      select(id, label, class, has_broader, description) %>%
+      left_join(inclConcepts %>% select(-description), by = c("id", "label", "class", "has_broader")) %>%
+      distinct(label, class, id, has_broader, .keep_all = TRUE) |>
       filter(!is.na(class))
 
     toJoin <- relate %>%
       rename(label_harm = label) %>%
       mutate(label = tolower(label_harm)) %>%
-      # filter(!is.na(label_harm) & class == tail(filterClasses, 1)) %>%
       filter(!is.na(label_harm)) %>%
+      filter(class == tail(filterClasses, 1)) |>
+      # filter(has_broader %in% parentFilter) |>
       select(-has_broader, -has_broader_match, -has_close_match, -has_exact_match, -has_narrower_match)
 
-    joined <- missingConcepts %>%
+    tempJoin <- missingConcepts %>%
       select(label_new = label, has_broader) %>%
-      mutate(label = tolower(label_new)) %>%
-      stringdist_left_join(toJoin, by = "label", distance_col = "dist", max_dist = 2)
+      mutate(label = tolower(label_new))
+
+    if(stringdist){
+      joined <- stringdist_left_join(x = tempJoin, y = toJoin, by = "label", distance_col = "dist", max_dist = 2) |>
+        separate_wider_regex(id, c(id_new = ".*", "[.]", rest = ".*"), cols_remove = FALSE) |>
+        filter(has_broader == id_new) |>
+        select(-id_new, -rest)
+    } else {
+      joined <- left_join(x = tempJoin, y = toJoin, by = "label") |>
+        separate_wider_regex(id, c(id_new = ".*", "[.]", rest = ".*"), cols_remove = FALSE) |>
+        filter(has_broader == id_new) |>
+        select(-label, -id_new, -rest) |>
+        mutate(label.x = NA, label.y = NA, dist = 0)
+    }
 
     if(!all(is.na(joined$dist))){
       joined <- joined %>%
         select(-label.x, -label.y) %>%
+        distinct() %>%
         arrange(dist) %>%
         mutate(dist = paste0("dist_", dist)) %>%
         pivot_wider(names_from = dist, values_from = label_harm)
+
       if(!"dist_0" %in% colnames(joined)){
         joined <- joined %>%
           add_column(dist_0 = NA_character_, .after = "description")
@@ -242,41 +313,36 @@ edit_matches <- function(new, target = NULL, source = NULL,
 
       hits <- joined %>%
         filter(!is.na(has_0_differences)) %>%
-        mutate(#class = tail(filterClasses, 1),
-               has_new_close_match = label,
+        mutate(has_new_close_match = label,
                label = has_0_differences) %>%
         select(label, id, all_of(withBroader), class, has_new_close_match)
 
-      numbers <- relate %>%
-        group_by(label) %>%
-        summarise(n = n())
-
       relate <- relate %>%
         left_join(hits, by = c("id", "label", "class", withBroader)) %>%
-        left_join(numbers, by = "label") %>%
-        mutate(has_new_close_match = if_else(n > 1, NA_character_, has_new_close_match)) %>%
-        mutate(has_new_close_match = if_else(str_detect(has_close_match, has_new_close_match), NA_character_, has_new_close_match)) %>%
+        rowwise() %>%
+        mutate(has_new_close_match = if_else(has_close_match %in% has_new_close_match, NA_character_, has_new_close_match)) %>%
         unite(col = "has_close_match", has_close_match, has_new_close_match, sep = " | ", na.rm = TRUE) %>%
-        mutate(across(where(is.character), function(x) na_if(x, ""))) %>%
-        select(-n)
+        mutate(across(where(is.character), function(x) na_if(x, "")))
 
-      missingJoined <- joined %>%
-        filter(is.na(has_0_differences))
       stillMissing <- joined %>%
+        filter(is.na(has_0_differences)) %>%
         select(-has_broader, -has_0_differences, -id, -class) %>%
         group_by(label) %>%
         summarise(across(starts_with("has_"), ~ paste0(na.omit(unique(.x)), collapse = " | "))) %>%
         mutate(across(where(is.character), function(x) na_if(x, ""))) %>%
         ungroup()
 
-      missingConcepts <- missingConcepts %>%
-        filter(label %in% missingJoined$label) %>%
+      stillMissing <- missingConcepts %>%
+        filter(!label %in% hits$has_new_close_match) %>%
         left_join(stillMissing, by = "label")
 
+    } else {
+      stillMissing <- missingConcepts
     }
 
-    sortIn <- missingConcepts %>%
-      left_join(tibble(label = new), by = "label") %>%
+    sortIn <- stillMissing %>%
+      # could also include the class of the stillMissing concepts to clarify
+      left_join(tibble(label = unique(new$label)), by = "label") %>%
       mutate(sort_in = label,
              label = NA_character_,
              class = NA_character_) %>%
@@ -293,19 +359,12 @@ edit_matches <- function(new, target = NULL, source = NULL,
         beep(sound = beep)
       }
 
-      # ... and make them aware of their duty
-      if(prevAvail){
-        message("\nprevious matches found for this dataseries, only previously not matched terms are presented")
-      } else {
-        message("\nno previous matches found for this dataseries, close match with other potentially available terms is presented")
-      }
-
-      message("-> please edit the file '", paste0(matchDir, "/matching.csv"), "' \n")
+      message("-> please edit the file '", paste0(matchDir, "/matching.csv"), "' (at class ", tail(filterClasses, 1), ") \n")
       if(verbose){
         message("--- column description ---\n")
         message("sort_in             cut out these values and sort them either into 'has_broader_match', \n                    'has_exact_match', has_narrower_match or 'has_close_match'")
-        message("has_broader         the broader concept id of each of the already harmonised concepts")
-        message("id                  filter by this column to jump to the subset you need to edit")
+        message("has_broader         the id of the broader concept of the already harmonised concepts")
+        message("id                  the id of concepts to which the new terms should be related")
         message("label               concepts to which the new terms should be related")
         message("class               the class of harmonised concepts")
         message("description         the description of each concept")
@@ -321,24 +380,30 @@ edit_matches <- function(new, target = NULL, source = NULL,
 
       related <- read_csv(paste0(matchDir, "/matching.csv"), col_types = cols(.default = "c"))
       assertNames(x = names(related), must.include = c("sort_in", "has_broader", "id", "label", "class", "description", "has_broader_match", "has_close_match", "has_exact_match", "has_narrower_match"))
-      related <- related %>%
-        select(-sort_in) %>%
-        filter(!is.na(id)) %>%
-        select(-any_of(c("has_0_differences", "has_1_difference", "has_2_differences")))
+      toIgnore <- related %>%
+        filter(id == "ignore") %>%
+        mutate(label = "ignore",
+               class = ignoreClass,
+               has_close_match = sort_in,
+               id = str_replace_all(ontology@classes$harmonised$id[ontology@classes$harmonised$label == ignoreClass], pattern = "x", replacement = "0")) %>%
+        group_by(id, label, class, has_broader_match, has_exact_match, has_narrower_match) %>%
+        summarise(has_close_match = paste0(na.omit(has_close_match), collapse = " | "), .groups = "keep") %>%
+        ungroup()
 
-      # if("dist" %in% names(joined)){
-      #   if(!all(is.na(joined$dist))){
-      #     related <- related %>%
-      #       select(-any_of("has_0_differences", "has_1_difference", "has_2_differences"))
-      #   }
-      # }
+      related <- related %>%
+        filter(!is.na(id) & id != "ignore") %>%
+        select(-sort_in) %>%
+        mutate(description = NA_character_) %>%
+        select(-any_of(c("has_0_differences", "has_1_difference", "has_2_differences"))) %>%
+        bind_rows(toIgnore)
 
       if(dim(related)[1] == 0){
         related <- NULL
       }
 
     } else {
-      related <- relate
+      related <- relate %>%
+        mutate(description = NA_character_)
     }
 
   } else {
@@ -347,26 +412,48 @@ edit_matches <- function(new, target = NULL, source = NULL,
 
   if(!is.null(related)){
 
-    out <- prevMatches %>%
+    matchingTable <- dsMatches %>%
       filter(!id == "ignore") %>%
       bind_rows(related) %>%
       pivot_longer(cols = c(has_broader_match, has_close_match, has_exact_match, has_narrower_match),
                    names_to = "match", values_to = "new_label") %>%
-      separate_rows(new_label, sep = " \\| ") %>%
+      separate_longer_delim(cols = new_label, delim = " | ") %>%
       distinct() %>%
-      group_by(id, has_broader, label, class, description, match) %>%
-      summarise(new_label = paste0(na.omit(new_label), collapse = " | "), .groups = "keep") %>%
-      ungroup() %>%
-      mutate(new_label = na_if(x = new_label, y = "")) %>%
-      pivot_wider(id_cols = c(label, class, id, has_broader, description), names_from = match, values_from = new_label) %>%
-      distinct() %>%
+      pivot_wider(id_cols = c("id", "label", "class", "has_broader", "description"), names_from = match,
+                  values_from = new_label, values_fn = ~paste0(na.omit(.x), collapse = " | ")) %>%
+      mutate(across(where(is.character), ~na_if(x = ., y = ""))) %>%
       filter(!is.na(has_broader_match) | !is.na(has_close_match) | !is.na(has_narrower_match) | !is.na(has_exact_match)) %>%
       filter(!is.na(id)) %>%
       arrange(id)
 
+    saveRDS(object = matchingTable, file = paste0(matchDir, sourceFile))
+
+    # only return objects that fall within those broader concepts that are in
+    # the final matching table at the current class
+    new_parentFilter <- matchingTable |>
+      filter(class == tail(filterClasses, 1)) |>
+      distinct(has_broader) |>
+      pull(has_broader)
+
+    if(!all(is.na(new_parentFilter))){
+      related <- related %>%
+        filter(has_broader %in% new_parentFilter)
+    }
+
+    newGrep <- str_replace_all(new$label, c("\\(" = "\\\\(", "\\)" = "\\\\)", "\\*" = "\\\\*",
+                                            "\\[" = "\\\\[", "\\]" = "\\\\]"))
+
+    out <- related %>%
+      pivot_longer(cols = c(has_broader_match, has_close_match, has_exact_match, has_narrower_match),
+                   names_to = "match", values_to = "external") %>%
+      separate_longer_delim(cols = external, delim = " | ") %>%
+      filter(!is.na(external)) |>
+      distinct() |>
+      mutate(match = str_replace(string = match, pattern = "has_", replacement = ""),
+             match = str_replace(string = match, pattern = "_match", replacement = "")) |>
+      filter(external %in% new$label)
+
   }
 
-  write_csv(x = out, file = paste0(matchDir, sourceFile), append = FALSE, na = "")
-
-  return(related)
+  return(out)
 }